refactor(core): split config/mod.rs test module into concern shards (#8280)
Extract the 3.2k-LOC `#[cfg(test)] mod tests` block from
crates/tsz-core/src/config/mod.rs into config/tests/{options_parsing,
module_resolution,strict_lib_extends}.rs, each well under the 2000-LOC
§19 limit. mod.rs drops 8206 -> 4981 lines and the #8280 arch-guard cap
is ratcheted to match in the same diff.

Pure code relocation (no production change): all 175 config tests pass
unchanged, clippy --tests clean, arch guard green.

AgentName: M1Opus48
