+++ crates/tsz-core/src/config/mod.rs
@@ -4978,3229 +4978,4 @@ fn find_lib_entry_offset(source: &str, entry: &str) -> u32 {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_parse_boolean_true() {
-        let json = r#"{"strict": true}"#;
-        let opts: CompilerOptions = serde_json::from_str(json).unwrap();
-        assert_eq!(opts.strict, Some(true));
-    }
-
-    #[test]
-    fn test_parse_string_true() {
-        let json = r#"{"strict": "true"}"#;
-        let opts: CompilerOptions = serde_json::from_str(json).unwrap();
-        assert_eq!(opts.strict, Some(true));
-    }
-
-    #[test]
-    fn test_parse_invalid_string() {
-        let json = r#"{"strict": "invalid"}"#;
-        let result: Result<CompilerOptions, _> = serde_json::from_str(json);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_esnext_date_and_temporal_are_valid_lib_values() {
-        // esnext.date and esnext.temporal must be in VALID_LIB_VALUES so tsconfig
-        // using "@lib: esnext.date,esnext.temporal" does not emit TS6046.
-        assert!(
-            VALID_LIB_VALUES.contains(&"esnext.date"),
-            "esnext.date should be a recognized lib value"
-        );
-        assert!(
-            VALID_LIB_VALUES.contains(&"esnext.temporal"),
-            "esnext.temporal should be a recognized lib value"
-        );
-    }
-
-    #[test]
-    fn test_esnext_date_temporal_lib_in_tsconfig_no_ts6046() {
-        let source = r#"{"compilerOptions":{"lib":["esnext.date","esnext.temporal"]}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6046),
-            "esnext.date and esnext.temporal should not emit TS6046, got: {codes:?}"
-        );
-    }
-
-    #[test]
-
-    fn test_ts5024_emitted_for_lib_replacement_string_value() {
-        let source = r#"{"compilerOptions":{"libReplacement":"true"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5024),
-            "Expected TS5024 for libReplacement string value, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5024_emitted_for_scalar_compiler_options() {
-        // Issue #3882: a top-level scalar `compilerOptions` would bypass every
-        // nested option validator and trip serde's `invalid type` error
-        // instead of the user-facing TS5024 diagnostic.
-        let source = r#"{"compilerOptions":"bad","files":["a.ts"]}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let diag = parsed
-            .diagnostics
-            .iter()
-            .find(|d| d.code == 5024 && d.message_text.contains("compilerOptions"))
-            .unwrap_or_else(|| {
-                panic!(
-                    "Expected TS5024 mentioning compilerOptions, got: {:?}",
-                    parsed
-                        .diagnostics
-                        .iter()
-                        .map(|d| (d.code, d.message_text.clone()))
-                        .collect::<Vec<_>>()
-                )
-            });
-        assert!(
-            diag.message_text.contains("object"),
-            "TS5024 message should mention expected type 'object': {}",
-            diag.message_text
-        );
-    }
-
-    #[test]
-    fn test_scalar_compiler_options_does_not_break_files_array() {
-        // The recovery path replaces the invalid scalar with `{}` so the
-        // rest of the config (e.g. `files`) still parses cleanly.
-        let source = r#"{"compilerOptions":42,"files":["a.ts"]}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert_eq!(
-            parsed.config.files.as_deref(),
-            Some(&["a.ts".to_string()][..]),
-            "files array must still parse after compilerOptions recovery"
-        );
-    }
-
-    #[test]
-    fn test_ts5024_emitted_for_non_boolean_compile_on_save() {
-        // Issue #3591 repro C: `compileOnSave` is a top-level boolean. A
-        // string value must surface as TS5024, matching tsc.
-        let source = r#"{
-  "compilerOptions": {"noEmit": true},
-  "compileOnSave": "yes",
-  "files": ["a.ts"]
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let diag = parsed
-            .diagnostics
-            .iter()
-            .find(|d| d.code == 5024 && d.message_text.contains("compileOnSave"))
-            .unwrap_or_else(|| {
-                panic!(
-                    "Expected TS5024 mentioning compileOnSave, got: {:?}",
-                    parsed
-                        .diagnostics
-                        .iter()
-                        .map(|d| (d.code, d.message_text.clone()))
-                        .collect::<Vec<_>>()
-                )
-            });
-        assert!(
-            diag.message_text.contains("boolean"),
-            "TS5024 message should mention expected type 'boolean': {}",
-            diag.message_text
-        );
-    }
-
-    #[test]
-    fn test_compile_on_save_boolean_value_does_not_emit_ts5024() {
-        // The validator must only fire for non-boolean values; a real boolean
-        // is accepted silently.
-        let source = r#"{
-  "compilerOptions": {"noEmit": true},
-  "compileOnSave": true,
-  "files": ["a.ts"]
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            !parsed
-                .diagnostics
-                .iter()
-                .any(|d| d.code == 5024 && d.message_text.contains("compileOnSave")),
-            "boolean compileOnSave must not emit TS5024, got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| (d.code, d.message_text.clone()))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts17010_emitted_for_unknown_type_acquisition_option() {
-        // Issue #3591 repro B: an unknown `typeAcquisition` key is silently
-        // dropped; tsc reports TS17010 for it.
-        let source = r#"{
-  "compilerOptions": {"noEmit": true},
-  "typeAcquisition": {"bogus": true},
-  "files": ["a.ts"]
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let diag = parsed
-            .diagnostics
-            .iter()
-            .find(|d| d.code == 17010 && d.message_text.contains("bogus"))
-            .unwrap_or_else(|| {
-                panic!(
-                    "Expected TS17010 mentioning bogus, got: {:?}",
-                    parsed
-                        .diagnostics
-                        .iter()
-                        .map(|d| (d.code, d.message_text.clone()))
-                        .collect::<Vec<_>>()
-                )
-            });
-        assert!(
-            diag.message_text
-                .to_lowercase()
-                .contains("type acquisition"),
-            "TS17010 message should reference 'type acquisition': {}",
-            diag.message_text
-        );
-    }
-
-    #[test]
-    fn test_known_type_acquisition_keys_do_not_emit_ts17010() {
-        // The four known typeAcquisition options must not be flagged.
-        let source = r#"{
-  "compilerOptions": {"noEmit": true},
-  "typeAcquisition": {
-    "enable": true,
-    "include": ["jquery"],
-    "exclude": ["lodash"],
-    "disableFilenameBasedTypeAcquisition": false
-  },
-  "files": ["a.ts"]
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let count = parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == 17010)
-            .count();
-        assert_eq!(
-            count,
-            0,
-            "no TS17010 expected for known typeAcquisition keys, got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| (d.code, d.message_text.clone()))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5024_emitted_for_scalar_type_acquisition() {
-        // A scalar typeAcquisition value (not an object) must surface as
-        // TS5024 via the shared object-shape validator.
-        let source = r#"{
-  "compilerOptions": {"noEmit": true},
-  "typeAcquisition": "bad",
-  "files": ["a.ts"]
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed
-                .diagnostics
-                .iter()
-                .any(|d| d.code == 5024 && d.message_text.contains("typeAcquisition")),
-            "expected TS5024 for scalar typeAcquisition, got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| (d.code, d.message_text.clone()))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5024_emitted_for_recognized_options_with_invalid_value_types() {
-        for (option, value, expected_type) in [
-            ("plugins", r#""not-an-array""#, "Array"),
-            ("maxNodeModuleJsDepth", r#""not-a-number""#, "number"),
-            ("traceResolution", r#""yes""#, "boolean"),
-        ] {
-            let source = format!(
-                r#"{{
-  "compilerOptions": {{
-    "noEmit": true,
-    "{option}": {value}
-  }},
-  "files": ["index.ts"]
-}}"#
-            );
-            let parsed =
-                parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap_or_else(|err| {
-                    panic!("{option} should report TS5024, not fail parse: {err}")
-                });
-            let diagnostic = parsed
-                .diagnostics
-                .iter()
-                .find(|d| d.code == 5024 && d.message_text.contains(option))
-                .unwrap_or_else(|| {
-                    panic!(
-                        "Expected TS5024 for invalid {option}, got: {:?}",
-                        parsed.diagnostics
-                    )
-                });
-            assert!(
-                diagnostic.message_text.contains(expected_type),
-                "Expected {option} TS5024 to mention {expected_type}, got: {diagnostic:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts5059_emitted_for_invalid_react_namespace_value() {
-        let source = r#"{
-  "compilerOptions": {
-    "jsx": "react",
-    "reactNamespace": "my-React-Lib"
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let ts5059 = parsed
-            .diagnostics
-            .iter()
-            .find(|d| {
-                d.code
-                    == diagnostic_codes::INVALID_VALUE_FOR_REACTNAMESPACE_IS_NOT_A_VALID_IDENTIFIER
-            })
-            .unwrap_or_else(|| panic!("Expected TS5059, got: {:?}", parsed.diagnostics));
-
-        assert_eq!(ts5059.file, "tsconfig.json");
-        assert_eq!(
-            ts5059.start,
-            source
-                .find("\"my-React-Lib\"")
-                .expect("reactNamespace value") as u32
-        );
-        assert!(
-            ts5059
-                .message_text
-                .contains("'my-React-Lib' is not a valid identifier"),
-            "Unexpected TS5059 message: {}",
-            ts5059.message_text
-        );
-    }
-
-    #[test]
-    fn test_disable_size_limit_option_is_recognized() {
-        let source = r#"{
-  "compilerOptions": {
-    "disableSizeLimit": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION),
-            "disableSizeLimit should not report unknown compiler option, got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN),
-            "disableSizeLimit should not report did-you-mean diagnostic, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_disable_solution_searching_option_is_recognized() {
-        let source = r#"{
-  "compilerOptions": {
-    "disableSolutionSearching": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION),
-            "disableSolutionSearching should not report unknown compiler option, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_tsz_only_compiler_options_report_unknown_from_tsconfig() {
-        let source = r#"{
-  "compilerOptions": {
-    "inlineConstants": true,
-    "disableSolutionTypeCheck": true,
-    "disableSolutionCaching": true,
-    "disableSolutionTypeChecking": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert_eq!(
-            codes.len(),
-            4,
-            "unrecognized options should produce tsc-compatible diagnostics, got: {:?}",
-            parsed.diagnostics
-        );
-        assert_eq!(
-            codes
-                .iter()
-                .filter(|&&code| code == diagnostic_codes::UNKNOWN_COMPILER_OPTION)
-                .count(),
-            2,
-            "expected two TS5023 diagnostics, got: {:?}",
-            parsed.diagnostics
-        );
-        assert_eq!(
-            codes
-                .iter()
-                .filter(|&&code| code == diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN)
-                .count(),
-            2,
-            "expected two TS5025 diagnostics, got: {:?}",
-            parsed.diagnostics
-        );
-        assert!(
-            parsed
-                .diagnostics
-                .iter()
-                .filter(|diag| diag.code == diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN)
-                .all(|diag| diag.message_text.contains("disableSolutionSearching")),
-            "disableSolution typo diagnostics should suggest disableSolutionSearching, got: {:?}",
-            parsed.diagnostics
-        );
-    }
-
-    #[test]
-    fn test_sound_tsconfig_option_enables_sound_mode() {
-        let source = r#"{"compilerOptions":{"sound":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.is_empty(),
-            "`sound` should be accepted as a known option, got: {:?}",
-            parsed.diagnostics
-        );
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            resolved.checker.sound_mode,
-            "sound: true in tsconfig should enable sound_mode"
-        );
-    }
-
-    #[test]
-    fn test_sound_tsconfig_option_false_keeps_sound_mode_off() {
-        let source = r#"{"compilerOptions":{"sound":false}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.is_empty(),
-            "`sound: false` should be accepted without diagnostics, got: {:?}",
-            parsed.diagnostics
-        );
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.sound_mode,
-            "false must not flip sound_mode on"
-        );
-    }
-
-    #[test]
-    fn test_sound_tsconfig_invalid_value_emits_ts5024() {
-        let source = r#"{"compilerOptions":{"sound":"yes_please"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&diagnostic_codes::COMPILER_OPTION_REQUIRES_A_VALUE_OF_TYPE),
-            "non-boolean `sound` value should emit TS5024, got: {:?}",
-            parsed.diagnostics
-        );
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.sound_mode,
-            "invalid `sound` value should not enable sound_mode"
-        );
-    }
-
-    #[test]
-    fn test_sound_family_tsconfig_options_accepted() {
-        let source = r#"{
-  "compilerOptions": {
-    "sound": true,
-    "soundCheckDeclarations": true,
-    "soundReportOnly": true,
-    "soundPedantic": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.is_empty(),
-            "sound family options should all be accepted, got: {:?}",
-            parsed.diagnostics
-        );
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            resolved.checker.sound_mode,
-            "sound: true should set sound_mode"
-        );
-        assert!(
-            resolved.checker.sound_check_declarations,
-            "soundCheckDeclarations: true should set sound_check_declarations"
-        );
-        assert!(
-            resolved.checker.sound_report_only,
-            "soundReportOnly: true should set sound_report_only"
-        );
-        assert!(
-            resolved.checker.sound_pedantic,
-            "soundPedantic: true should set sound_pedantic"
-        );
-    }
-
-    #[test]
-    fn test_sound_family_tsconfig_options_miscased_emit_did_you_mean() {
-        // All-uppercase-suffix spellings (e.g. soundPEDANTIC) still fall within
-        // the Levenshtein threshold for getSpellingSuggestion, so they get TS5025
-        // rather than a bare TS5023.
-        let cases = [
-            ("Sound", "sound"),
-            ("soundCheckdeclarations", "soundCheckDeclarations"),
-            ("soundreportonly", "soundReportOnly"),
-            ("soundPEDANTIC", "soundPedantic"),
-        ];
-        for (typo, canonical) in cases {
-            let source = format!(r#"{{"compilerOptions":{{"{typo}":true}}}}"#);
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let diag = parsed
-                .diagnostics
-                .iter()
-                .find(|d| d.code == diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN);
-            assert!(
-                diag.is_some(),
-                "typo `{typo}` should emit TS5025, got: {:?}",
-                parsed.diagnostics
-            );
-            assert!(
-                diag.unwrap().message_text.contains(canonical),
-                "TS5025 for `{typo}` should suggest `{canonical}`, got: {}",
-                diag.unwrap().message_text
-            );
-        }
-    }
-
-    #[test]
-    fn test_disable_size_limit_miscase_reports_did_you_mean() {
-        let source = r#"{
-  "compilerOptions": {
-    "disableSizelimit": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN),
-            "Expected TS5025-style did-you-mean for mis-cased disableSizeLimit, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_typo_suggestions_emit_ts5025_for_close_compiler_option_names() {
-        // Issue #3831: typoed compiler-option names that the canonical option
-        // list contains within Levenshtein range should be reported as TS5025
-        // with a `Did you mean ...` suggestion, matching tsc.
-        let cases = [
-            ("stric", "strict"),
-            ("noEmti", "noEmit"),
-            ("moduleResoluton", "moduleResolution"),
-        ];
-        for (typo, expected) in cases {
-            let source = format!("{{\"compilerOptions\":{{\"{typo}\":true}}}}");
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let diag = parsed
-                .diagnostics
-                .iter()
-                .find(|d| d.code == diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN)
-                .unwrap_or_else(|| {
-                    panic!(
-                        "Expected TS5025 for typo {typo}, got: {:?}",
-                        parsed.diagnostics
-                    )
-                });
-            assert!(
-                diag.message_text.contains(expected),
-                "TS5025 for {typo} should suggest {expected}, got: {}",
-                diag.message_text
-            );
-            assert!(
-                !parsed
-                    .diagnostics
-                    .iter()
-                    .any(|d| d.code == diagnostic_codes::UNKNOWN_COMPILER_OPTION),
-                "Bare TS5023 should not be emitted alongside TS5025 for {typo}, got: {:?}",
-                parsed.diagnostics
-            );
-        }
-    }
-
-    #[test]
-    fn test_unrelated_unknown_compiler_option_still_falls_back_to_ts5023() {
-        // A name that bears no resemblance to any canonical option must not
-        // get a spurious TS5025 suggestion, just the bare TS5023.
-        let source = r#"{"compilerOptions":{"completelyUnrelatedXYZ":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION),
-            "Expected TS5023 for unrelated unknown option, got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN),
-            "Unrelated option should not emit a Did-you-mean suggestion, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn explain_files_compiler_option_is_recognized() {
-        // Issue #3860: `explainFiles` was missing from the known-options
-        // list, causing tsconfig-side `explainFiles: true` to surface a
-        // false TS5023 even though tsc accepts it from config.
-        let source = r#"{"compilerOptions":{"explainFiles":true,"noEmit":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION),
-            "explainFiles must not emit TS5023, got {codes:?}"
-        );
-    }
-
-    #[test]
-    fn explain_files_lowercase_alias_is_recognized() {
-        // Miscased `explainfiles` should map to `explainFiles` and emit
-        // TS5025 (Did you mean), not the bare TS5023.
-        let source = r#"{"compilerOptions":{"explainfiles":true,"noEmit":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION_DID_YOU_MEAN),
-            "miscased explainfiles must emit TS5025 Did you mean, got {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_disable_source_of_project_reference_redirect_option_is_recognized() {
-        let source = r#"{
-  "compilerOptions": {
-    "disableSourceOfProjectReferenceRedirect": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&diagnostic_codes::UNKNOWN_COMPILER_OPTION),
-            "disableSourceOfProjectReferenceRedirect should not report unknown compiler option, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_resolve_compiler_options_sets_lib_replacement_flag() {
-        let json = r#"{"compilerOptions":{"libReplacement":true}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(resolved.lib_replacement);
-    }
-
-    #[test]
-    fn test_tsconfig_emit_flags_reach_printer_options() {
-        let json = r#"{"compilerOptions":{"importHelpers":true,"preserveConstEnums":true,"downlevelIteration":true}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-
-        assert!(resolved.import_helpers);
-        assert!(resolved.printer.import_helpers);
-        assert!(resolved.printer.no_emit_helpers);
-        assert!(resolved.checker.preserve_const_enums);
-        assert!(resolved.printer.preserve_const_enums);
-        assert!(resolved.checker.downlevel_iteration);
-        assert!(resolved.printer.downlevel_iteration);
-    }
-
-    #[test]
-    fn test_parse_module_resolution_rejects_comma_separated_value() {
-        let json =
-            r#"{"compilerOptions":{"moduleResolution":"node16,nodenext","module":"commonjs"}} "#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let err = resolve_compiler_options(config.compiler_options.as_ref())
-            .expect_err("comma-separated moduleResolution should be rejected");
-        assert!(
-            err.to_string().contains("compilerOptions.moduleResolution"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn test_ts6046_emitted_for_comma_separated_enum_options() {
-        for (option, value, flag) in [
-            ("target", "es2020,esnext", "--target"),
-            ("module", "commonjs,esnext", "--module"),
-            ("moduleResolution", "node,bundler", "--moduleResolution"),
-            ("moduleDetection", "auto,force", "--moduleDetection"),
-            ("newLine", "lf,crlf", "--newLine"),
-        ] {
-            let source = format!(r#"{{"compilerOptions":{{"{option}":"{value}"}}}}"#);
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let diagnostic = parsed
-                .diagnostics
-                .iter()
-                .find(|diag| diag.code == diagnostic_codes::ARGUMENT_FOR_OPTION_MUST_BE)
-                .unwrap_or_else(|| {
-                    panic!(
-                        "Expected TS6046 for compilerOptions.{option}, got: {:?}",
-                        parsed.diagnostics
-                    )
-                });
-
-            assert!(
-                diagnostic.message_text.contains(flag),
-                "Unexpected TS6046 message for compilerOptions.{option}: {}",
-                diagnostic.message_text
-            );
-            assert_eq!(
-                diagnostic.start,
-                source.find(&format!(r#""{value}""#)).unwrap() as u32
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts6046_emitted_for_separator_mutated_enum_options() {
-        for (option, value, flag) in [
-            ("target", "es_2020", "--target"),
-            ("target", "es-2020", "--target"),
-            ("target", "es 2020", "--target"),
-            ("module", "node_next", "--module"),
-            ("jsx", "react_jsx", "--jsx"),
-            ("moduleResolution", "node_16", "--moduleResolution"),
-        ] {
-            let source = format!(
-                r#"{{"compilerOptions":{{"{option}":"{value}","noEmit":true}},"files":["a.ts"]}}"#
-            );
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let codes: Vec<u32> = parsed.diagnostics.iter().map(|diag| diag.code).collect();
-            let diagnostic = parsed
-                .diagnostics
-                .iter()
-                .find(|diag| diag.code == diagnostic_codes::ARGUMENT_FOR_OPTION_MUST_BE)
-                .unwrap_or_else(|| {
-                    panic!(
-                        "Expected TS6046 for compilerOptions.{option}={value:?}, got: {:?}",
-                        parsed.diagnostics
-                    )
-                });
-
-            assert!(
-                diagnostic.message_text.contains(flag),
-                "Unexpected TS6046 message for compilerOptions.{option}: {}",
-                diagnostic.message_text
-            );
-            assert!(
-                !codes.contains(
-                    &diagnostic_codes::OPTION_MODULE_MUST_BE_SET_TO_WHEN_OPTION_MODULERESOLUTION_IS_SET_TO
-                ),
-                "separator-mutated moduleResolution should not produce follow-on TS5110, got: {:?}",
-                parsed.diagnostics
-            );
-            assert_eq!(
-                diagnostic.start,
-                source.find(&format!(r#""{value}""#)).unwrap() as u32
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts6046_emitted_for_invalid_module_detection_and_new_line() {
-        for (option, value, flag, expected_values) in [
-            (
-                "moduleDetection",
-                "bogus",
-                "--moduleDetection",
-                "'auto', 'legacy', 'force'",
-            ),
-            ("newLine", "bogus", "--newLine", "'crlf', 'lf'"),
-        ] {
-            let source = format!(r#"{{"compilerOptions":{{"{option}":"{value}"}}}}"#);
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let diagnostic = parsed
-                .diagnostics
-                .iter()
-                .find(|diag| diag.code == diagnostic_codes::ARGUMENT_FOR_OPTION_MUST_BE)
-                .unwrap_or_else(|| {
-                    panic!(
-                        "Expected TS6046 for compilerOptions.{option}, got: {:?}",
-                        parsed.diagnostics
-                    )
-                });
-
-            assert!(
-                diagnostic.message_text.contains(flag)
-                    && diagnostic.message_text.contains(expected_values),
-                "Unexpected TS6046 message for compilerOptions.{option}: {}",
-                diagnostic.message_text
-            );
-            assert_eq!(
-                diagnostic.start,
-                source.find(&format!(r#""{value}""#)).unwrap() as u32
-            );
-
-            let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref())
-                .expect("invalid enum value should be nulled before resolution");
-            if option == "moduleDetection" {
-                assert!(!resolved.printer.module_detection_force);
-                assert!(!resolved.printer.module_detection_legacy);
-            }
-        }
-    }
-
-    #[test]
-    fn test_shared_module_defaults_cover_targets_and_resolution() {
-        assert_eq!(
-            default_module_kind_for_target(ScriptTarget::ES5, true),
-            ModuleKind::CommonJS
-        );
-        assert_eq!(
-            default_module_kind_for_target(ScriptTarget::ES2019, true),
-            ModuleKind::ES2015
-        );
-        assert_eq!(
-            default_module_kind_for_target(ScriptTarget::ES2021, true),
-            ModuleKind::ES2020
-        );
-        assert_eq!(
-            default_module_kind_for_target(ScriptTarget::ES2025, true),
-            ModuleKind::ES2022
-        );
-        assert_eq!(
-            default_module_kind_for_target(ScriptTarget::ES2025, false),
-            ModuleKind::ESNext
-        );
-        assert_eq!(
-            default_module_resolution_for_module(ModuleKind::System),
-            ModuleResolutionKind::Classic
-        );
-        assert_eq!(
-            default_module_resolution_for_module(ModuleKind::CommonJS),
-            ModuleResolutionKind::Bundler
-        );
-        assert_eq!(
-            default_module_resolution_for_module(ModuleKind::Node20),
-            ModuleResolutionKind::Node16
-        );
-        assert_eq!(
-            default_module_resolution_for_module(ModuleKind::NodeNext),
-            ModuleResolutionKind::NodeNext
-        );
-    }
-
-    #[test]
-    fn test_module_explicitly_set_when_specified() {
-        let json = r#"{"compilerOptions":{"module":"es2015"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(resolved.checker.module_explicitly_set);
-        assert!(resolved.checker.module.is_es_module());
-    }
-
-    #[test]
-    fn test_module_explicitly_set_commonjs() {
-        let json = r#"{"compilerOptions":{"module":"commonjs"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(resolved.checker.module_explicitly_set);
-        assert!(!resolved.checker.module.is_es_module());
-    }
-
-    #[test]
-    fn test_module_not_explicitly_set_defaults_from_target() {
-        // When module is not specified, it's computed from target.
-        // module_explicitly_set is false (module was derived, not explicit).
-        let json = r#"{"compilerOptions":{"target":"es2015"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(!resolved.checker.module_explicitly_set);
-        // Module defaults to ES2015 for es2015+ targets
-        assert!(resolved.checker.module.is_es_module());
-    }
-
-    #[test]
-    fn test_effective_module_resolution_defaults_to_bundler_for_es_modules() {
-        // tsc 6.0: ES module kinds default to Bundler resolution (was Classic)
-        let json = r#"{"compilerOptions":{"module":"es2015","target":"es2015"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert_eq!(
-            resolved.effective_module_resolution(),
-            ModuleResolutionKind::Bundler
-        );
-    }
-
-    #[test]
-    fn test_no_config_defaults_to_bundler_and_resolve_json_module() {
-        let resolved = resolve_compiler_options(None).unwrap();
-
-        assert_eq!(
-            resolved.effective_module_resolution(),
-            ModuleResolutionKind::Bundler
-        );
-        assert!(resolved.resolve_json_module);
-        assert!(resolved.checker.resolve_json_module);
-    }
-
-    #[test]
-    fn test_effective_module_resolution_prefers_explicit_override() {
-        let json = r#"{"compilerOptions":{"module":"es2015","moduleResolution":"bundler","target":"es2015"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert_eq!(
-            resolved.effective_module_resolution(),
-            ModuleResolutionKind::Bundler
-        );
-    }
-
-    #[test]
-    fn test_module_not_explicitly_set_no_options() {
-        // When no options at all, module_explicitly_set should be false.
-        let resolved = resolve_compiler_options(None).unwrap();
-        assert!(!resolved.checker.module_explicitly_set);
-        assert!(
-            resolved.printer.always_strict,
-            "printer alwaysStrict should default to true with no compiler options"
-        );
-    }
-
-    #[test]
-    fn test_removed_compiler_option_lookup() {
-        assert!(removed_compiler_option("noImplicitUseStrict").is_some());
-        assert!(removed_compiler_option("keyofStringsOnly").is_some());
-        assert!(removed_compiler_option("suppressExcessPropertyErrors").is_some());
-        assert!(removed_compiler_option("suppressImplicitAnyIndexErrors").is_some());
-        assert!(removed_compiler_option("noStrictGenericChecks").is_some());
-        assert!(removed_compiler_option("charset").is_some());
-        assert!(removed_compiler_option("out").is_some());
-        assert_eq!(
-            removed_compiler_option("importsNotUsedAsValues"),
-            Some("verbatimModuleSyntax")
-        );
-        assert_eq!(
-            removed_compiler_option("preserveValueImports"),
-            Some("verbatimModuleSyntax")
-        );
-        // Non-removed options return None
-        assert!(removed_compiler_option("strict").is_none());
-        assert!(removed_compiler_option("target").is_none());
-    }
-
-    #[test]
-    fn test_ts5102_emitted_for_removed_option() {
-        let source = r#"{"compilerOptions":{"noImplicitUseStrict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5102),
-            "Expected TS5102 for removed option noImplicitUseStrict, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_not_emitted_for_false_removed_option() {
-        // When a removed boolean option is set to false, tsc doesn't emit TS5102
-        let source = r#"{"compilerOptions":{"noImplicitUseStrict":false}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5102),
-            "Should NOT emit TS5102 for false-valued removed option, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_emitted_for_string_removed_option() {
-        let source = r#"{"compilerOptions":{"importsNotUsedAsValues":"error"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5102),
-            "Expected TS5102 for removed option importsNotUsedAsValues, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_not_suppressed_with_ignore_deprecations() {
-        // In tsc 6.0, removed options (deprecated 5.0, removed 5.5) always emit TS5102
-        // because mustBeRemoved is true (removedIn 5.5 <= tsc 6.0).
-        // ignoreDeprecations only suppresses TS5101 (deprecated but not yet removed).
-        let source =
-            r#"{"compilerOptions":{"ignoreDeprecations":"5.0","noImplicitUseStrict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5102),
-            "Should emit TS5102 even with ignoreDeprecations '5.0' (option is past removal), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_not_suppressed_with_invalid_ignore_deprecations() {
-        // Invalid ignoreDeprecations value should NOT suppress TS5102
-        let source =
-            r#"{"compilerOptions":{"ignoreDeprecations":"7.0","noImplicitUseStrict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5102),
-            "Should emit TS5102 when ignoreDeprecations is invalid, got: {codes:?}"
-        );
-        assert!(
-            codes.contains(&5103),
-            "Should also emit TS5103 for invalid ignoreDeprecations, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_fires_with_ignore_deprecations_6_0() {
-        // "6.0" IS a valid ignoreDeprecations value in tsc 6.0.
-        // TS5102 still fires for removed 5.0-wave options (past removal deadline).
-        // TS5103 must NOT fire because "6.0" is valid.
-        let source =
-            r#"{"compilerOptions":{"ignoreDeprecations":"6.0","noImplicitUseStrict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5102),
-            "Should emit TS5102 even with ignoreDeprecations '6.0' (option is past removal), got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&5103),
-            "Should NOT emit TS5103 — '6.0' is a valid ignoreDeprecations value, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_fires_for_all_removed_options() {
-        // Verify all removed options trigger TS5102 unconditionally
-        let removed_opts = [
-            ("noImplicitUseStrict", "true"),
-            ("keyofStringsOnly", "true"),
-            ("suppressExcessPropertyErrors", "true"),
-            ("suppressImplicitAnyIndexErrors", "true"),
-            ("noStrictGenericChecks", "true"),
-            ("charset", r#""utf8""#),
-            ("importsNotUsedAsValues", r#""error""#),
-            ("preserveValueImports", "true"),
-            ("out", r#""out.js""#),
-        ];
-        for (opt, val) in &removed_opts {
-            let source =
-                format!(r#"{{"compilerOptions":{{"{opt}":{val},"ignoreDeprecations":"6.0"}}}}"#);
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-            assert!(
-                codes.contains(&5102),
-                "Should emit TS5102 for removed option '{opt}' even with ignoreDeprecations '6.0', got: {codes:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts5102_inherited_from_extends_anchors_at_child_compiler_options_key() {
-        // Repro from `verbatimModuleSyntaxCompat3.ts`. When the extending
-        // tsconfig.json uses `verbatimModuleSyntax` and the base tsconfig
-        // contains removed options (`preserveValueImports`,
-        // `importsNotUsedAsValues`), tsc anchors TS5102 at the *child's*
-        // `"compilerOptions"` key — not at `"verbatimModuleSyntax"` which
-        // tsz used to incorrectly anchor on. Reproducing requires real
-        // tempfiles because the inheritance resolution reads from disk.
-        use tempfile::tempdir;
-        let temp = tempdir().expect("create temp dir");
-        let base_path = temp.path().join("tsconfig.base.json");
-        let child_path = temp.path().join("tsconfig.json");
-        std::fs::write(
-            &base_path,
-            r#"{
-    "compilerOptions": {
-        "isolatedModules": true,
-        "preserveValueImports": true,
-        "importsNotUsedAsValues": "error"
-    }
-}"#,
-        )
-        .expect("write base");
-        let child_source = r#"{
-    "extends": "./tsconfig.base.json",
-    "compilerOptions": {
-        "verbatimModuleSyntax": true
-    }
-}"#;
-        std::fs::write(&child_path, child_source).expect("write child");
-
-        let parsed = load_tsconfig_with_diagnostics(&child_path).expect("load");
-        let ts5102: Vec<&Diagnostic> = parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == 5102)
-            .collect();
-        assert!(
-            ts5102.len() >= 2,
-            "Expected at least 2 TS5102 (preserveValueImports + importsNotUsedAsValues), got: {ts5102:?}"
-        );
-        // Each TS5102 must anchor at the child's `"compilerOptions"` key.
-        let expected_start = child_source
-            .find("\"compilerOptions\"")
-            .expect("compilerOptions in child source") as u32;
-        for diag in &ts5102 {
-            assert_eq!(
-                diag.start, expected_start,
-                "Inherited TS5102 must anchor at child's `\"compilerOptions\"` key (start={expected_start}), got: {diag:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_inherited_base_url_anchored_at_base_config_dir() {
-        // tsc resolves a tsconfig's `baseUrl` relative to the config file
-        // that declares it. When a child extends a base that sets
-        // `baseUrl: "."`, the inherited `baseUrl` must point at the *base*
-        // config's directory, not the child's. Issue #3332 reproduced the
-        // child-anchored bug, which broke inherited `paths` mappings.
-        let temp = tempdir().expect("create temp dir");
-        let base_dir = temp.path().join("base");
-        let app_dir = temp.path().join("app");
-        std::fs::create_dir_all(&base_dir).expect("create base dir");
-        std::fs::create_dir_all(&app_dir).expect("create app dir");
-
-        let base_path = base_dir.join("tsconfig.base.json");
-        std::fs::write(
-            &base_path,
-            r#"{
-    "compilerOptions": {
-        "baseUrl": ".",
-        "paths": { "@shared/*": ["shared/*"] }
-    }
-}"#,
-        )
-        .expect("write base");
-
-        let child_path = app_dir.join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{
-    "extends": "../base/tsconfig.base.json",
-    "files": ["src/index.ts"]
-}"#,
-        )
-        .expect("write child");
-
-        let merged = load_tsconfig(&child_path).expect("load child");
-        let opts = merged.compiler_options.expect("compiler options merged");
-        let base_url = opts.base_url.expect("inherited baseUrl present");
-
-        // Canonicalize to handle macOS `/var` → `/private/var` symlinks.
-        let canonical_base_dir = std::fs::canonicalize(&base_dir).unwrap_or(base_dir);
-        let canonical_app_dir = std::fs::canonicalize(&app_dir).unwrap_or(app_dir);
-        let canonical_base_url = std::path::Path::new(&base_url)
-            .canonicalize()
-            .unwrap_or_else(|_| std::path::PathBuf::from(&base_url));
-        let expected = canonical_base_dir.to_string_lossy();
-        let actual = canonical_base_url.to_string_lossy();
-        assert!(
-            actual.starts_with(expected.as_ref()),
-            "Inherited baseUrl must anchor at the base config's directory \
-             (expected prefix {expected:?}, got {actual:?})"
-        );
-        assert!(
-            !actual.starts_with(canonical_app_dir.to_string_lossy().as_ref()),
-            "Inherited baseUrl must not anchor at the child's directory: {actual:?}"
-        );
-    }
-
-    #[test]
-    fn test_inherited_root_dirs_anchor_at_declaring_config_dir() {
-        let temp = tempdir().expect("create temp dir");
-        let base_dir = temp.path().join("base");
-        let app_dir = temp.path().join("app");
-        std::fs::create_dir_all(&base_dir).expect("create base dir");
-        std::fs::create_dir_all(&app_dir).expect("create app dir");
-
-        let base_path = base_dir.join("tsconfig.base.json");
-        std::fs::write(
-            &base_path,
-            r#"{
-    "compilerOptions": {
-        "rootDirs": ["src", "generated"]
-    }
-}"#,
-        )
-        .expect("write base");
-
-        let child_path = app_dir.join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{
-    "extends": "../base/tsconfig.base.json",
-    "files": ["src/index.ts"]
-}"#,
-        )
-        .expect("write child");
-
-        let merged = load_tsconfig(&child_path).expect("load child");
-        let opts = merged.compiler_options.expect("compiler options merged");
-        let root_dirs = opts.root_dirs.expect("inherited rootDirs present");
-        let expected_base = base_dir
-            .canonicalize()
-            .expect("canonicalize base")
-            .to_string_lossy()
-            .into_owned();
-        let unexpected_app = app_dir
-            .canonicalize()
-            .expect("canonicalize app")
-            .to_string_lossy()
-            .into_owned();
-
-        assert_eq!(root_dirs.len(), 2);
-        for root_dir in &root_dirs {
-            assert!(
-                root_dir.starts_with(&expected_base),
-                "Inherited rootDirs must anchor at the base config's directory, got {root_dir:?}"
-            );
-            assert!(
-                !root_dir.starts_with(&unexpected_app),
-                "Inherited rootDirs must not anchor at the child's directory: {root_dir:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_inherited_path_options_anchor_at_declaring_config_dir() {
-        let temp = tempdir().expect("create temp dir");
-        let base_dir = temp.path().join("base");
-        let app_dir = temp.path().join("app");
-        std::fs::create_dir_all(&base_dir).expect("create base dir");
-        std::fs::create_dir_all(&app_dir).expect("create app dir");
-
-        let base_path = base_dir.join("tsconfig.base.json");
-        std::fs::write(
-            &base_path,
-            r#"{
-    "compilerOptions": {
-        "rootDir": "src",
-        "outDir": "dist",
-        "declarationDir": "types",
-        "tsBuildInfoFile": ".cache/project.tsbuildinfo",
-        "typeRoots": ["./types"]
-    }
-}"#,
-        )
-        .expect("write base");
-
-        let child_path = app_dir.join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{
-    "extends": "../base/tsconfig.base.json",
-    "files": ["src/index.ts"]
-}"#,
-        )
-        .expect("write child");
-
-        let merged = load_tsconfig(&child_path).expect("load child");
-        let opts = merged.compiler_options.expect("compiler options merged");
-        let type_roots = opts.type_roots.expect("inherited typeRoots present");
-        let expected_base = base_dir
-            .canonicalize()
-            .expect("canonicalize base")
-            .to_string_lossy()
-            .into_owned();
-        let unexpected_app = app_dir
-            .canonicalize()
-            .expect("canonicalize app")
-            .to_string_lossy()
-            .into_owned();
-
-        for (name, value) in [
-            ("rootDir", opts.root_dir.expect("rootDir present")),
-            ("outDir", opts.out_dir.expect("outDir present")),
-            (
-                "declarationDir",
-                opts.declaration_dir.expect("declarationDir present"),
-            ),
-            (
-                "tsBuildInfoFile",
-                opts.ts_build_info_file.expect("tsBuildInfoFile present"),
-            ),
-        ] {
-            assert!(
-                value.starts_with(&expected_base),
-                "Inherited {name} must anchor at the base config's directory, got {value:?}"
-            );
-            assert!(
-                !value.starts_with(&unexpected_app),
-                "Inherited {name} must not anchor at the child's directory: {value:?}"
-            );
-        }
-
-        assert_eq!(type_roots.len(), 1);
-        let type_root = &type_roots[0];
-        assert!(
-            type_root.starts_with(&expected_base),
-            "Inherited typeRoots must anchor at the base config's directory, got {type_root:?}"
-        );
-        assert!(
-            !type_root.starts_with(&unexpected_app),
-            "Inherited typeRoots must not anchor at the child's directory: {type_root:?}"
-        );
-    }
-
-    #[test]
-    fn test_child_base_url_overrides_inherited_and_anchors_at_child_dir() {
-        // When the child config also declares `baseUrl`, the child wins
-        // and is resolved relative to the child's directory (matching tsc).
-        let temp = tempdir().expect("create temp dir");
-        let base_dir = temp.path().join("base");
-        let app_dir = temp.path().join("app");
-        std::fs::create_dir_all(&base_dir).expect("create base dir");
-        std::fs::create_dir_all(&app_dir).expect("create app dir");
-
-        let base_path = base_dir.join("tsconfig.base.json");
-        std::fs::write(&base_path, r#"{ "compilerOptions": { "baseUrl": "." } }"#)
-            .expect("write base");
-
-        let child_path = app_dir.join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{
-    "extends": "../base/tsconfig.base.json",
-    "compilerOptions": { "baseUrl": "src" }
-}"#,
-        )
-        .expect("write child");
-
-        let merged = load_tsconfig(&child_path).expect("load child");
-        let opts = merged.compiler_options.expect("compiler options merged");
-        let base_url = opts.base_url.expect("baseUrl present");
-
-        // Canonicalize both sides so symlink-bearing temp paths on macOS
-        // (`/var/folders/...` → `/private/var/folders/...`) compare equal.
-        let canonical_app_dir = std::fs::canonicalize(&app_dir).unwrap_or(app_dir);
-        let canonical_base_url = std::path::Path::new(&base_url)
-            .canonicalize()
-            .unwrap_or_else(|_| std::path::PathBuf::from(&base_url));
-        let expected_prefix = canonical_app_dir.to_string_lossy();
-        let actual = canonical_base_url.to_string_lossy();
-        assert!(
-            actual.starts_with(expected_prefix.as_ref()),
-            "Child-declared baseUrl must anchor at the child's directory \
-             (expected prefix {expected_prefix:?}, got {actual:?})"
-        );
-    }
-
-    #[test]
-    fn test_inherited_absolute_base_url_is_preserved() {
-        // An absolute `baseUrl` declared in the base config must propagate
-        // unchanged through `extends`.
-        let temp = tempdir().expect("create temp dir");
-        let base_dir = temp.path().join("base");
-        let app_dir = temp.path().join("app");
-        let abs_base_url = temp.path().join("shared-root");
-        std::fs::create_dir_all(&base_dir).expect("create base dir");
-        std::fs::create_dir_all(&app_dir).expect("create app dir");
-        std::fs::create_dir_all(&abs_base_url).expect("create shared root");
-
-        let abs_str = abs_base_url.to_string_lossy().replace('\\', "/");
-        let base_path = base_dir.join("tsconfig.base.json");
-        std::fs::write(
-            &base_path,
-            format!(r#"{{ "compilerOptions": {{ "baseUrl": "{abs_str}" }} }}"#),
-        )
-        .expect("write base");
-
-        let child_path = app_dir.join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{ "extends": "../base/tsconfig.base.json" }"#,
-        )
-        .expect("write child");
-
-        let merged = load_tsconfig(&child_path).expect("load child");
-        let base_url = merged
-            .compiler_options
-            .expect("compiler options merged")
-            .base_url
-            .expect("baseUrl present");
-        assert_eq!(
-            std::path::Path::new(&base_url),
-            abs_base_url.as_path(),
-            "Absolute inherited baseUrl must be preserved verbatim"
-        );
-    }
-
-    #[test]
-    fn test_ts5102_not_emitted_for_valid_option() {
-        let source = r#"{"compilerOptions":{"strict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5102),
-            "Should NOT emit TS5102 for valid option 'strict', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_not_emitted_for_bundler_with_commonjs() {
-        // tsc 6.0 allows moduleResolution: bundler with module: commonjs
-        let source = r#"{"compilerOptions":{"module":"commonjs","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5095),
-            "Should NOT emit TS5095 for bundler+commonjs, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_bundler_with_none() {
-        let source = r#"{"compilerOptions":{"module":"none","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5095),
-            "Expected TS5095 for bundler+none, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_bundler_with_amd() {
-        let source = r#"{"compilerOptions":{"module":"amd","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5095),
-            "Expected TS5095 for bundler+amd, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_bundler_with_system() {
-        let source = r#"{"compilerOptions":{"module":"system","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5095),
-            "Expected TS5095 for bundler+system, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_not_emitted_for_bundler_with_es2015() {
-        let source = r#"{"compilerOptions":{"module":"es2015","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5095),
-            "Should NOT emit TS5095 for bundler+es2015, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_not_emitted_for_bundler_with_esnext() {
-        let source = r#"{"compilerOptions":{"module":"esnext","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5095),
-            "Should NOT emit TS5095 for bundler+esnext, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_not_emitted_for_bundler_with_preserve() {
-        let source = r#"{"compilerOptions":{"module":"preserve","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5095),
-            "Should NOT emit TS5095 for bundler+preserve, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_emitted_for_bundler_with_node16() {
-        let source = r#"{"compilerOptions":{"module":"node16","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5095),
-            "Should emit TS5095 for bundler+node16 (tsc behavior), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_emitted_for_bundler_with_node18() {
-        let source = r#"{"compilerOptions":{"module":"node18","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5095),
-            "Should emit TS5095 for bundler+node18 (tsc behavior), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_emitted_for_bundler_with_nodenext() {
-        let source = r#"{"compilerOptions":{"module":"nodenext","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5095),
-            "Should emit TS5095 for bundler+nodenext (tsc behavior), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5095_not_emitted_for_node16_resolution() {
-        let source = r#"{"compilerOptions":{"module":"commonjs","moduleResolution":"node16"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5095),
-            "Should NOT emit TS5095 for node16 resolution, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_emitted_for_invalid_ignore_deprecations() {
-        // tsz conservatively emits TS5103 whenever ignoreDeprecations is set to an invalid value.
-        // tsc only emits TS5103 when deprecated features are also present, but since tsz cannot
-        // detect all deprecated features (e.g. deprecated source syntax like import assertions),
-        // it conservatively emits TS5103 for any invalid ignoreDeprecations value.
-        let source = r#"{"compilerOptions":{"ignoreDeprecations":"7.0"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5103),
-            "Expected TS5103 for invalid ignoreDeprecations='7.0', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_emitted_for_invalid_ignore_deprecations_with_deprecated_option() {
-        // tsc emits TS5103 when an invalid ignoreDeprecations value is used alongside
-        // a removed/deprecated option (the invalid value can't suppress the warning).
-        let source =
-            r#"{"compilerOptions":{"ignoreDeprecations":"5.1","noImplicitUseStrict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5103),
-            "Expected TS5103 for ignoreDeprecations='5.1' with deprecated option, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_emitted_for_invalid_ignore_deprecations_with_deprecated_target_alias() {
-        // tsc emits TS5103 when an invalid ignoreDeprecations value is used alongside
-        // a deprecated target alias like "es6" (deprecated in favor of "es2015").
-        // This matches the arrowFunction conformance test pattern.
-        let source = r#"{"compilerOptions":{"ignoreDeprecations":"7.0","target":"es6"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5103),
-            "Expected TS5103 for ignoreDeprecations='7.0' with deprecated target='es6', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_emitted_for_invalid_ignore_deprecations_with_any_target() {
-        // tsz emits TS5103 conservatively for any invalid ignoreDeprecations value,
-        // regardless of target. Even non-deprecated targets like "es2018" will trigger
-        // TS5103 in tsz (conservative approach since we can't detect all deprecated syntax).
-        let source = r#"{"compilerOptions":{"ignoreDeprecations":"7.0","target":"es2018"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5103),
-            "Expected TS5103 (conservative) for ignoreDeprecations='7.0' with target='es2018', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_not_emitted_for_valid_value() {
-        let source = r#"{"compilerOptions":{"ignoreDeprecations":"5.0"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5103),
-            "Should NOT emit TS5103 for valid ignoreDeprecations='5.0', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_not_emitted_for_valid_6_0() {
-        // tsc 6.0 accepts both "5.0" and "6.0" as valid ignoreDeprecations values.
-        // See TypeScript/src/compiler/program.ts getIgnoreDeprecationsVersion():
-        //   if (ignoreDeprecations === "5.0" || ignoreDeprecations === "6.0") return new Version(...)
-        let source = r#"{"compilerOptions":{"ignoreDeprecations":"6.0"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5103),
-            "Should NOT emit TS5103 for valid ignoreDeprecations='6.0', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_not_emitted_for_6_0_with_deprecated_options() {
-        // ignoreDeprecations: "6.0" silences 6.0-wave deprecation warnings.
-        // TS5102 still fires for removed 5.0-wave options (noImplicitUseStrict is removed),
-        // but TS5103 must NOT fire because "6.0" is a valid value.
-        let source =
-            r#"{"compilerOptions":{"ignoreDeprecations":"6.0","noImplicitUseStrict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5102),
-            "Should still emit TS5102 for removed option, got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&5103),
-            "Should NOT emit TS5103 — '6.0' is a valid ignoreDeprecations value, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_emitted_for_invalid_5_5() {
-        // tsc 6.0 only accepts "5.0" — "5.5" is not a valid ignoreDeprecations value
-        let source = r#"{"compilerOptions":{"ignoreDeprecations":"5.5"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5103),
-            "Should emit TS5103 for invalid ignoreDeprecations='5.5', got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5103_not_emitted_when_absent() {
-        let source = r#"{"compilerOptions":{"strict":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5103),
-            "Should NOT emit TS5103 when ignoreDeprecations is absent, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5110_node16_resolution_with_commonjs_module() {
-        let source = r#"{"compilerOptions":{"module":"commonjs","moduleResolution":"node16"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5110),
-            "Should emit TS5110 for node16 resolution with commonjs module, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5110_nodenext_resolution_with_es2022_module() {
-        let source = r#"{"compilerOptions":{"module":"es2022","moduleResolution":"nodenext"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5110),
-            "Should emit TS5110 for nodenext resolution with es2022 module, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5110_not_emitted_for_matching_node16() {
-        let source = r#"{"compilerOptions":{"module":"node16","moduleResolution":"node16"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5110),
-            "Should NOT emit TS5110 when module matches moduleResolution, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5110_not_emitted_for_matching_nodenext() {
-        let source = r#"{"compilerOptions":{"module":"nodenext","moduleResolution":"nodenext"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5110),
-            "Should NOT emit TS5110 when module matches moduleResolution, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5069_emit_declaration_only_without_declaration() {
-        let source = r#"{"compilerOptions":{"emitDeclarationOnly":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5069),
-            "Expected TS5069 for emitDeclarationOnly without declaration, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5069_not_emitted_with_declaration() {
-        let source = r#"{"compilerOptions":{"emitDeclarationOnly":true,"declaration":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5069),
-            "Should NOT emit TS5069 when declaration is true, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5069_not_emitted_with_composite() {
-        let source = r#"{"compilerOptions":{"emitDeclarationOnly":true,"composite":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5069),
-            "Should NOT emit TS5069 when composite is true, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5069_emitted_when_declaration_has_string_boolean() {
-        let source = r#"{
-  "compilerOptions": {
-    "declaration": "true",
-    "emitDeclarationOnly": true
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let ts5024_count = parsed.diagnostics.iter().filter(|d| d.code == 5024).count();
-        assert_eq!(
-            ts5024_count, 1,
-            "Expected TS5024 for string-typed declaration"
-        );
-
-        let mut ts5069_starts: Vec<u32> = parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == 5069)
-            .map(|d| d.start)
-            .collect();
-        ts5069_starts.sort_unstable();
-        assert_eq!(
-            ts5069_starts.len(),
-            2,
-            "Expected TS5069 at both declaration and emitDeclarationOnly"
-        );
-        assert_eq!(
-            ts5069_starts,
-            vec![
-                find_key_offset_in_source(source, "declaration"),
-                find_key_offset_in_source(source, "emitDeclarationOnly"),
-            ]
-        );
-    }
-
-    #[test]
-    fn test_ts5069_declaration_map_without_declaration() {
-        let source = r#"{"compilerOptions":{"declarationMap":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5069),
-            "Expected TS5069 for declarationMap without declaration, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5053_sourcemap_with_inline_sourcemap() {
-        let source = r#"{"compilerOptions":{"sourceMap":true,"inlineSourceMap":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5053),
-            "Expected TS5053 for sourceMap with inlineSourceMap, got: {codes:?}"
-        );
-        // tsc emits twice (at each key position)
-        let count = codes.iter().filter(|&&c| c == 5053).count();
-        assert_eq!(
-            count, 2,
-            "Expected 2 TS5053 diagnostics (one per key), got: {count}"
-        );
-    }
-
-    #[test]
-    fn test_ts5053_not_emitted_without_conflict() {
-        let source = r#"{"compilerOptions":{"sourceMap":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5053),
-            "Should NOT emit TS5053 for sourceMap alone, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5053_allow_js_with_isolated_declarations() {
-        let source = r#"{"compilerOptions":{"allowJs":true,"isolatedDeclarations":true,"declaration":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5053),
-            "Expected TS5053 for allowJs with isolatedDeclarations, got: {codes:?}"
-        );
-    }
-
-    // Issue #3732: when checkJs is true and allowJs is absent, tsc treats
-    // allowJs as implied-true and still emits TS5053 for the
-    // (allowJs, isolatedDeclarations) conflict.
-    #[test]
-    fn test_ts5053_check_js_implies_allow_js_with_isolated_declarations() {
-        let source = r#"{"compilerOptions":{"checkJs":true,"isolatedDeclarations":true,"declaration":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5053),
-            "Expected TS5053 when checkJs implies allowJs alongside isolatedDeclarations, got: {codes:?}"
-        );
-        // The conflict message should still reference allowJs (the option
-        // tsc reports as conflicting), even though the diagnostic anchors
-        // at checkJs.
-        let ts5053: Vec<_> = parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == 5053)
-            .collect();
-        assert!(
-            ts5053.iter().any(|d| d.message_text.contains("'allowJs'")),
-            "Expected TS5053 message to reference allowJs, got: {ts5053:?}"
-        );
-    }
-
-    // Sanity: explicit `allowJs: false` must not implicitly enable allowJs
-    // through checkJs, so TS5053 must NOT fire.
-    #[test]
-    fn test_ts5053_check_js_with_explicit_allow_js_false_does_not_fire() {
-        let source = r#"{"compilerOptions":{"checkJs":true,"allowJs":false,"isolatedDeclarations":true,"declaration":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5053),
-            "Should not emit TS5053 when allowJs is explicitly false, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5052_not_emitted_when_check_js_implies_allow_js() {
-        let source = r#"{"compilerOptions":{"checkJs":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let has_5052 = parsed.diagnostics.iter().any(|d| d.code == 5052);
-        assert!(
-            !has_5052,
-            "Should not emit TS5052 when checkJs implies allowJs, got: {:?}",
-            parsed.diagnostics
-        );
-    }
-
-    #[test]
-    fn test_ts5052_check_js_with_allow_js_false_reports_both_sites() {
-        let source = r#"{"compilerOptions":{"allowJs":false,"checkJs":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let count = parsed.diagnostics.iter().filter(|d| d.code == 5052).count();
-        assert_eq!(
-            count, 2,
-            "Expected two TS5052 diagnostics (allowJs/checkJs), got: {:?}",
-            parsed.diagnostics
-        );
-    }
-
-    #[test]
-    fn test_ts5052_not_emitted_when_check_js_and_allow_js_true() {
-        let source = r#"{"compilerOptions":{"allowJs":true,"checkJs":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let has_5052 = parsed.diagnostics.iter().any(|d| d.code == 5052);
-        assert!(
-            !has_5052,
-            "Should not emit TS5052 when allowJs is true, got: {:?}",
-            parsed.diagnostics
-        );
-    }
-
-    #[test]
-    fn test_resolve_compiler_options_propagates_check_js_to_checker_options() {
-        let source = r#"{"compilerOptions":{"allowJs":true,"checkJs":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-
-        assert!(resolved.check_js);
-        assert!(resolved.checker.check_js);
-    }
-
-    #[test]
-    fn test_resolve_compiler_options_check_js_implies_allow_js() {
-        let source = r#"{"compilerOptions":{"checkJs":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-
-        assert!(resolved.check_js);
-        assert!(resolved.checker.check_js);
-        assert!(resolved.allow_js);
-        assert!(resolved.checker.allow_js);
-    }
-
-    #[test]
-    fn test_ts5070_resolve_json_module_with_classic_module_resolution() {
-        let source =
-            r#"{"compilerOptions":{"resolveJsonModule":true,"moduleResolution":"classic"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5070),
-            "Expected TS5070 for resolveJsonModule with classic moduleResolution, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_resolve_json_module_not_implied_by_node_resolution() {
-        for source in [
-            r#"{"compilerOptions":{"module":"commonjs","moduleResolution":"node10"}}"#,
-            r#"{"compilerOptions":{"module":"node16","moduleResolution":"node16"}}"#,
-        ] {
-            let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-            let resolved =
-                resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-
-            assert!(
-                !resolved.resolve_json_module,
-                "resolveJsonModule should not be implied for {source}"
-            );
-            assert!(
-                !resolved.checker.resolve_json_module,
-                "checker resolveJsonModule should not be implied for {source}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_resolve_json_module_implied_by_bundler_resolution() {
-        let source = r#"{"compilerOptions":{"moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-
-        assert!(resolved.resolve_json_module);
-        assert!(resolved.checker.resolve_json_module);
-    }
-
-    #[test]
-    fn test_ts5070_resolve_json_module_with_amd_module() {
-        // module=amd defaults to moduleResolution=classic
-        let source = r#"{"compilerOptions":{"resolveJsonModule":true,"module":"amd"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5070),
-            "Expected TS5070 for resolveJsonModule with module=amd (implies classic), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_resolve_json_module_with_system_module() {
-        // module=system without explicit moduleResolution implies classic resolution →
-        // tsc emits TS5070 (not TS5071) because the moduleResolution-based check takes precedence.
-        let source = r#"{"compilerOptions":{"resolveJsonModule":true,"module":"system"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5070),
-            "Expected TS5070 (not TS5071) for resolveJsonModule with module=system (implies classic), got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&5071),
-            "Should NOT emit TS5071 when effective moduleResolution is classic (TS5070 takes precedence), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_resolve_json_module_with_system_module_explicit_resolution() {
-        // module=system with explicit non-classic moduleResolution → TS5071 fires
-        let source = r#"{"compilerOptions":{"resolveJsonModule":true,"module":"system","moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5071),
-            "Expected TS5071 for resolveJsonModule with module=system + moduleResolution=bundler, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_resolve_json_module_with_none_module() {
-        // module=none without explicit moduleResolution implies classic resolution →
-        // tsc emits TS5070 (not TS5071) because the moduleResolution-based check takes precedence.
-        let source = r#"{"compilerOptions":{"resolveJsonModule":true,"module":"none"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5070),
-            "Expected TS5070 (not TS5071) for resolveJsonModule with module=none (implies classic), got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&5071),
-            "Should NOT emit TS5071 when effective moduleResolution is classic (TS5070 takes precedence), got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5098_resolve_package_json_with_classic() {
-        let source = r#"{"compilerOptions":{"resolvePackageJsonExports":true,"moduleResolution":"classic"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5098),
-            "Expected TS5098 for resolvePackageJsonExports with classic moduleResolution, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5098_not_emitted_with_bundler() {
-        let source = r#"{"compilerOptions":{"resolvePackageJsonExports":true,"moduleResolution":"bundler"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5098),
-            "Should NOT emit TS5098 with bundler moduleResolution, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5098_not_emitted_when_module_and_resolution_omitted() {
-        // #3509: tsz used to emit TS5098 for `customConditions` /
-        // `resolvePackageJsonExports` / `resolvePackageJsonImports` when
-        // both `module` and `moduleResolution` were unset, even though
-        // tsz's own defaulting chain (target=ESNext → module=ESNext →
-        // moduleResolution=Bundler) would land on a "modern" mode. tsc
-        // accepts the same configs.
-        for opt in [
-            "customConditions",
-            "resolvePackageJsonExports",
-            "resolvePackageJsonImports",
-        ] {
-            let source = if opt == "customConditions" {
-                format!(r#"{{"compilerOptions":{{"{opt}":["x"]}},"files":["index.ts"]}}"#)
-            } else {
-                format!(r#"{{"compilerOptions":{{"{opt}":true}},"files":["index.ts"]}}"#)
-            };
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-            assert!(
-                !codes.contains(&5098),
-                "must not emit TS5098 for {opt} when module/moduleResolution omitted, got {codes:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts5098_emitted_with_explicit_classic() {
-        // Explicit `moduleResolution: "classic"` must still trigger TS5098 —
-        // user opted out of the modern defaulting chain.
-        let source =
-            r#"{"compilerOptions":{"customConditions":["x"],"moduleResolution":"classic"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5098),
-            "explicit classic moduleResolution must still emit TS5098, got {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_resolve_extends_path_uses_package_exports_mapping() {
-        let temp = tempdir().unwrap();
-        let project_dir = temp.path().join("project");
-        let package_dir = project_dir.join("node_modules").join("pkg");
-        let config_dir = package_dir.join("configs");
-
-        std::fs::create_dir_all(&config_dir).unwrap();
-        std::fs::write(project_dir.join("tsconfig.json"), "{}").unwrap();
-        std::fs::write(
-            package_dir.join("package.json"),
-            r#"{
-                "exports": {
-                    "./tsconfig.json": "./configs/tsconfig.base.json"
-                }
-            }"#,
-        )
-        .unwrap();
-        let expected = config_dir.join("tsconfig.base.json");
-        std::fs::write(&expected, "{}").unwrap();
-
-        let resolved =
-            resolve_extends_path(&project_dir.join("tsconfig.json"), "pkg/tsconfig.json").unwrap();
-
-        assert_eq!(resolved, expected);
-    }
-
-    #[test]
-    fn test_ts6082_outfile_with_commonjs() {
-        let source = r#"{"compilerOptions":{"module":"commonjs","outFile":"all.js"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&6082),
-            "Expected TS6082 for outFile+commonjs, got: {codes:?}"
-        );
-        // Should emit twice — once at "module" key, once at "outFile" key
-        let count = codes.iter().filter(|&&c| c == 6082).count();
-        assert_eq!(
-            count, 2,
-            "Expected two TS6082 diagnostics (module + outFile keys), got {count}"
-        );
-    }
-
-    #[test]
-    fn test_ts6082_outfile_with_umd() {
-        let source = r#"{"compilerOptions":{"module":"umd","outFile":"all.js"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&6082),
-            "Expected TS6082 for outFile+umd, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6082_outfile_with_es6() {
-        let source = r#"{"compilerOptions":{"module":"es6","outFile":"all.js"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&6082),
-            "Expected TS6082 for outFile+es6, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6082_not_emitted_for_amd() {
-        let source = r#"{"compilerOptions":{"module":"amd","outFile":"all.js"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6082),
-            "Should NOT emit TS6082 for outFile+amd, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6082_not_emitted_for_system() {
-        let source = r#"{"compilerOptions":{"module":"system","outFile":"all.js"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6082),
-            "Should NOT emit TS6082 for outFile+system, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6082_not_emitted_with_emit_declaration_only() {
-        let source = r#"{"compilerOptions":{"module":"commonjs","outFile":"all.js","emitDeclarationOnly":true,"declaration":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6082),
-            "Should NOT emit TS6082 when emitDeclarationOnly is true, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6082_not_emitted_without_outfile() {
-        let source = r#"{"compilerOptions":{"module":"commonjs"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6082),
-            "Should NOT emit TS6082 without outFile, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_bundler_implied_resolve_json_module_with_umd() {
-        // moduleResolution: bundler implies resolveJsonModule=true.
-        // Combined with module=umd, this should emit TS5071.
-        let source = r#"{"compilerOptions":{"moduleResolution":"bundler","module":"umd"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5071),
-            "Expected TS5071 for bundler-implied resolveJsonModule with module=umd, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_bundler_implied_resolve_json_module_with_system() {
-        let source = r#"{"compilerOptions":{"moduleResolution":"bundler","module":"system"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&5071),
-            "Expected TS5071 for bundler-implied resolveJsonModule with module=system, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_explicit_resolve_json_module_reports_both_keys() {
-        let source = r#"{"compilerOptions":{"module":"none","moduleResolution":"bundler","resolveJsonModule":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let ts5071: Vec<_> = parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == 5071)
-            .collect();
-        assert_eq!(
-            ts5071.len(),
-            2,
-            "Expected TS5071 at both 'module' and 'resolveJsonModule', got: {ts5071:?}"
-        );
-        let starts: Vec<u32> = ts5071.iter().map(|d| d.start).collect();
-        assert!(
-            starts.contains(&find_key_offset_in_source(source, "module")),
-            "Expected TS5071 anchored to 'module', got starts: {starts:?}"
-        );
-        assert!(
-            starts.contains(&find_key_offset_in_source(source, "resolveJsonModule")),
-            "Expected TS5071 anchored to 'resolveJsonModule', got starts: {starts:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts5071_not_emitted_for_bundler_with_esnext() {
-        // moduleResolution: bundler + module=esnext should NOT emit TS5071
-        let source = r#"{"compilerOptions":{"moduleResolution":"bundler","module":"esnext"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5071),
-            "Should NOT emit TS5071 for bundler+esnext, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_implied_classic_resolution_es2015_module() {
-        // ES module kinds now default to Bundler resolution, so Classic should
-        // remain disabled here.
-        let json = r#"{"compilerOptions":{"module":"es2015"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.implied_classic_resolution,
-            "ES2015 should not imply Classic resolution"
-        );
-    }
-
-    #[test]
-    fn test_implied_classic_resolution_amd_module() {
-        let json = r#"{"compilerOptions":{"module":"amd"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            resolved.checker.implied_classic_resolution,
-            "AMD should imply Classic resolution"
-        );
-    }
-
-    #[test]
-    fn test_implied_classic_resolution_explicit_node_override() {
-        // module: es2015 + moduleResolution: node10 → NOT Classic
-        let json = r#"{"compilerOptions":{"module":"es2015","moduleResolution":"node10"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.implied_classic_resolution,
-            "Explicit moduleResolution: node10 should override Classic inference"
-        );
-    }
-
-    #[test]
-    fn test_implied_classic_resolution_commonjs_module() {
-        // module: commonjs → effective resolution is Node10, NOT Classic
-        let json = r#"{"compilerOptions":{"module":"commonjs"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.implied_classic_resolution,
-            "CommonJS module should not imply Classic resolution"
-        );
-    }
-
-    #[test]
-    fn test_implied_classic_resolution_nodenext_module() {
-        // module: nodenext → effective resolution is NodeNext, NOT Classic
-        let json = r#"{"compilerOptions":{"module":"nodenext"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.implied_classic_resolution,
-            "NodeNext module should not imply Classic resolution"
-        );
-    }
-
-    #[test]
-    fn test_implied_classic_resolution_explicit_bundler() {
-        // module: esnext + moduleResolution: bundler → NOT Classic
-        let json = r#"{"compilerOptions":{"module":"esnext","moduleResolution":"bundler"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.implied_classic_resolution,
-            "Explicit moduleResolution: bundler should override Classic inference"
-        );
-    }
-
-    #[test]
-    fn test_preserve_symlinks_defaults_false() {
-        let resolved = resolve_compiler_options(None).unwrap();
-        assert!(!resolved.preserve_symlinks);
-    }
-
-    #[test]
-    fn test_preserve_symlinks_resolved_from_tsconfig() {
-        let json = r#"{"compilerOptions":{"preserveSymlinks":true}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(resolved.preserve_symlinks);
-    }
-
-    #[test]
-    fn test_ts5107_always_strict_false() {
-        let source = r#"{"compilerOptions":{"alwaysStrict":false}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.iter().any(|d| d.code == 5107),
-            "alwaysStrict=false should trigger TS5107; got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| d.code)
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5107_target_es5() {
-        let source = r#"{"compilerOptions":{"target":"ES5"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.iter().any(|d| d.code == 5107),
-            "target=ES5 should trigger TS5107; got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| d.code)
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5108_target_es3() {
-        // target=ES3 was removed in TS 6.0 (TS5108, not suppressible by ignoreDeprecations).
-        for value in &["ES3", "es3"] {
-            let source = format!(r#"{{"compilerOptions":{{"target":"{value}"}}}}"#);
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            assert!(
-                parsed.diagnostics.iter().any(|d| d.code == 5108),
-                "target={value} should trigger TS5108; got: {:?}",
-                parsed
-                    .diagnostics
-                    .iter()
-                    .map(|d| d.code)
-                    .collect::<Vec<_>>()
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts5108_target_es3_not_suppressed_by_ignore_deprecations() {
-        // TS5108 (removed value) must fire even when ignoreDeprecations="6.0" is set.
-        let source = r#"{"compilerOptions":{"target":"ES3","ignoreDeprecations":"6.0"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.iter().any(|d| d.code == 5108),
-            "ignoreDeprecations=6.0 must NOT suppress TS5108 (removed value); got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| d.code)
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5107_suppressed_by_ignore_deprecations_6_0() {
-        let source = r#"{"compilerOptions":{"alwaysStrict":false,"ignoreDeprecations":"6.0"}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            !parsed.diagnostics.iter().any(|d| d.code == 5107),
-            "ignoreDeprecations=6.0 should suppress TS5107; got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| d.code)
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_child_ignore_deprecations_suppresses_inherited_ts5107() {
-        let temp = tempdir().expect("create temp dir");
-        let base_path = temp.path().join("base.json");
-        std::fs::write(
-            &base_path,
-            r#"{
-  "compilerOptions": {
-    "moduleResolution": "node"
-  }
-}"#,
-        )
-        .expect("write base");
-
-        let child_path = temp.path().join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{
-  "extends": "./base.json",
-  "compilerOptions": {
-    "ignoreDeprecations": "6.0"
-  },
-  "files": ["a.ts"]
-}"#,
-        )
-        .expect("write child");
-
-        let parsed = load_tsconfig_with_diagnostics(&child_path).expect("load child");
-        assert!(
-            !parsed.diagnostics.iter().any(|d| d.code == 5107),
-            "child ignoreDeprecations=6.0 should suppress inherited TS5107, got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| (&d.file, d.code, &d.message_text))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5101_base_url() {
-        let source = r#"{"compilerOptions":{"baseUrl":"."}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.iter().any(|d| d.code == 5101),
-            "baseUrl should trigger TS5101; got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| d.code)
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5090_paths_substitutions_require_base_url() {
-        let source = r#"{
-  "compilerOptions": {
-    "paths": {
-      "@app/*": ["src/*", "./ok/*", "../up/*", "*"]
-    }
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let ts5090: Vec<_> = parsed
-            .diagnostics
-            .iter()
-            .filter(|diag| diag.code == 5090)
-            .collect();
-        assert_eq!(
-            ts5090.len(),
-            2,
-            "Expected TS5090 only for non-relative substitutions without baseUrl, got: {:?}",
-            parsed
-                .diagnostics
-                .iter()
-                .map(|d| (d.code, &d.message_text))
-                .collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn test_ts5090_paths_substitutions_suppressed_when_base_url_present() {
-        let source = r#"{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "@app/*": ["src/*", "*"]
-    }
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&5090),
-            "baseUrl should suppress TS5090 for non-relative substitutions, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_strict_family_defaults_true_when_strict_not_set() {
-        // tsc 6.0 defaults: strict-family options are true when not explicitly set.
-        // The tsc cache was generated with tsc 6.0-dev which has strict=true as its
-        // effective default. CheckerOptions::default() reflects this.
-        let json = r#"{"compilerOptions":{"target":"es2015"}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            resolved.checker.strict_null_checks,
-            "strictNullChecks should default to true when strict is not set"
-        );
-        assert!(
-            resolved.checker.strict_function_types,
-            "strictFunctionTypes should default to true when strict is not set"
-        );
-        assert!(
-            resolved.checker.no_implicit_any,
-            "noImplicitAny should default to true when strict is not set"
-        );
-        assert!(
-            resolved.checker.strict_property_initialization,
-            "strictPropertyInitialization should default to true when strict is not set"
-        );
-        assert!(
-            resolved.checker.no_implicit_this,
-            "noImplicitThis should default to true when strict is not set"
-        );
-        assert!(
-            resolved.checker.use_unknown_in_catch_variables,
-            "useUnknownInCatchVariables should default to true when strict is not set"
-        );
-    }
-
-    #[test]
-    fn test_resolve_compiler_options_propagates_no_check() {
-        let json = r#"{"compilerOptions":{"noCheck":true}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(resolved.no_check, "noCheck should be read from tsconfig");
-    }
-
-    #[test]
-    fn test_strict_false_keeps_always_strict_default() {
-        // In TS 6.0, strict:false still leaves alwaysStrict on by default unless it
-        // is explicitly set to false.
-        let json = r#"{"compilerOptions":{"strict":false}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.strict_null_checks,
-            "strictNullChecks should be false when strict: false"
-        );
-        assert!(
-            !resolved.checker.no_implicit_any,
-            "noImplicitAny should be false when strict: false"
-        );
-        assert!(
-            !resolved.checker.strict_property_initialization,
-            "strictPropertyInitialization should be false when strict: false"
-        );
-        assert!(
-            resolved.checker.always_strict,
-            "alwaysStrict should remain true by default when strict: false"
-        );
-        assert!(
-            resolved.printer.always_strict,
-            "printer alwaysStrict should remain true by default when strict: false"
-        );
-    }
-
-    #[test]
-    fn test_individual_strict_option_overrides_default() {
-        // Individual strict-family options should override the default.
-        let json = r#"{"compilerOptions":{"strictNullChecks":false}}"#;
-        let config: TsConfig = serde_json::from_str(json).unwrap();
-        let resolved = resolve_compiler_options(config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.strict_null_checks,
-            "strictNullChecks should be false when explicitly set to false"
-        );
-        // Other strict-family options should still be true (from defaults)
-        assert!(
-            resolved.checker.no_implicit_any,
-            "noImplicitAny should remain true when only strictNullChecks is overridden"
-        );
-    }
-
-    #[test]
-    fn test_ts5024_boolean_string_uses_always_strict_default() {
-        // tsc still enforces strict-mode syntax here: the invalid string value is
-        // rejected with TS5024, then option resolution falls back to the TS 6.0
-        // alwaysStrict default of true.
-        let source = r#"{
-  "compilerOptions": {
-    "strict": false,
-    "alwaysStrict": "true"
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        // TS5024 should be emitted for the string-typed boolean
-        let has_ts5024 = parsed.diagnostics.iter().any(|d| d.code == 5024);
-        assert!(
-            has_ts5024,
-            "Should emit TS5024 for string 'true' on boolean option"
-        );
-        // The invalid value itself is not applied, but alwaysStrict still falls
-        // back to its TS 6.0 default of true.
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            resolved.checker.always_strict,
-            "alwaysStrict should fall back to the TS 6.0 default when provided as a string-typed boolean"
-        );
-        assert!(
-            resolved.printer.always_strict,
-            "printer alwaysStrict should fall back to the TS 6.0 default when provided as a string-typed boolean"
-        );
-    }
-
-    #[test]
-    fn test_explicit_always_strict_false_overrides_default_even_with_strict_false() {
-        let source = r#"{
-  "compilerOptions": {
-    "strict": false,
-    "alwaysStrict": false,
-    "ignoreDeprecations": "6.0"
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.always_strict,
-            "explicit alwaysStrict=false should still disable alwaysStrict"
-        );
-        assert!(
-            !resolved.printer.always_strict,
-            "explicit alwaysStrict=false should still disable printer alwaysStrict"
-        );
-    }
-
-    #[test]
-    fn test_ts5024_isolated_modules_string_is_not_applied() {
-        let source = r#"{
-  "compilerOptions": {
-    "target": "es2015",
-    "module": "commonjs",
-    "isolatedModules": "true"
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let has_ts5024 = parsed.diagnostics.iter().any(|d| d.code == 5024);
-        assert!(
-            has_ts5024,
-            "Should emit TS5024 for string 'true' on isolatedModules"
-        );
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.checker.isolated_modules,
-            "isolatedModules should not be applied from a string-typed boolean value"
-        );
-    }
-
-    #[test]
-    fn test_ts5024_allow_importing_ts_extensions_string_is_not_applied() {
-        let source = r#"{
-  "compilerOptions": {
-    "moduleResolution": "bundler",
-    "module": "esnext",
-    "allowImportingTsExtensions": "true"
-  }
-}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let has_ts5024 = parsed.diagnostics.iter().any(|d| d.code == 5024);
-        assert!(
-            has_ts5024,
-            "Should emit TS5024 for string 'true' on allowImportingTsExtensions"
-        );
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            !resolved.allow_importing_ts_extensions,
-            "allowImportingTsExtensions should not be applied from a string-typed boolean value"
-        );
-    }
-
-    #[test]
-    fn test_ts5096_allow_importing_ts_extensions_requires_emit_guard() {
-        let invalid = r#"{"compilerOptions":{"allowImportingTsExtensions":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(invalid, "tsconfig.json").unwrap();
-        assert!(
-            parsed.diagnostics.iter().any(|d| d.code
-                == diagnostic_codes::OPTION_ALLOWIMPORTINGTSEXTENSIONS_CAN_ONLY_BE_USED_WHEN_ONE_OF_NOEMIT_EMITDECLAR),
-            "Expected TS5096 for allowImportingTsExtensions without an emit guard, got: {:?}",
-            parsed.diagnostics
-        );
-
-        for valid in [
-            r#"{"compilerOptions":{"allowImportingTsExtensions":true,"noEmit":true}}"#,
-            r#"{"compilerOptions":{"allowImportingTsExtensions":true,"declaration":true,"emitDeclarationOnly":true}}"#,
-            r#"{"compilerOptions":{"allowImportingTsExtensions":true,"rewriteRelativeImportExtensions":true}}"#,
-        ] {
-            let parsed = parse_tsconfig_with_diagnostics(valid, "tsconfig.json").unwrap();
-            assert!(
-                parsed.diagnostics.iter().all(|d| d.code
-                    != diagnostic_codes::OPTION_ALLOWIMPORTINGTSEXTENSIONS_CAN_ONLY_BE_USED_WHEN_ONE_OF_NOEMIT_EMITDECLAR),
-                "Did not expect TS5096 for guarded allowImportingTsExtensions in {valid}, got: {:?}",
-                parsed.diagnostics
-            );
-        }
-    }
-
-    #[test]
-    fn test_ts6304_composite_disables_declaration() {
-        let source = r#"{"compilerOptions":{"composite":true,"declaration":false}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&6304),
-            "Expected TS6304 when composite:true but declaration:false, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6304_not_emitted_when_declaration_true() {
-        let source = r#"{"compilerOptions":{"composite":true,"declaration":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6304),
-            "Should NOT emit TS6304 when both composite and declaration are true, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6379_composite_disables_incremental() {
-        let source = r#"{"compilerOptions":{"composite":true,"incremental":false}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&6379),
-            "Expected TS6379 when composite:true but incremental:false, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_ts6379_not_emitted_when_incremental_omitted() {
-        // composite implies incremental, so omitting incremental is fine
-        let source = r#"{"compilerOptions":{"composite":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let codes: Vec<u32> = parsed.diagnostics.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&6379),
-            "Should NOT emit TS6379 when composite is true and incremental is omitted, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn test_composite_implies_declaration_and_incremental() {
-        let source = r#"{"compilerOptions":{"composite":true,"noLib":true}}"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-        assert!(
-            resolved.composite,
-            "composite should be true in resolved options"
-        );
-        assert!(
-            resolved.emit_declarations,
-            "composite should imply declaration:true"
-        );
-        assert!(
-            resolved.incremental,
-            "composite should imply incremental:true"
-        );
-    }
-
-    #[test]
-    fn test_no_property_access_from_index_signature_resolves_from_tsconfig() {
-        let source = r#"{
-            "compilerOptions": {
-                "noPropertyAccessFromIndexSignature": true
-            }
-        }"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let resolved = resolve_compiler_options(parsed.config.compiler_options.as_ref()).unwrap();
-
-        assert!(resolved.checker.no_property_access_from_index_signature);
-    }
-
-    #[test]
-    fn test_ts5024_top_level_selector_type_mismatches_are_recovered() {
-        for (key, value) in [
-            ("include", r#""*.ts""#),
-            ("exclude", r#""dist""#),
-            ("references", r#""./lib""#),
-        ] {
-            let source = format!(
-                r#"{{
-  "{key}": {value},
-  "compilerOptions": {{ "noEmit": true }},
-  "files": ["a.ts"]
-}}"#
-            );
-            let parsed = parse_tsconfig_with_diagnostics(&source, "tsconfig.json").unwrap();
-            let ts5024 = parsed
-                .diagnostics
-                .iter()
-                .find(|d| {
-                    d.code == diagnostic_codes::COMPILER_OPTION_REQUIRES_A_VALUE_OF_TYPE
-                        && d.message_text.contains(key)
-                        && d.message_text.contains("Array")
-                })
-                .unwrap_or_else(|| {
-                    panic!("Expected TS5024 for {key}, got: {:?}", parsed.diagnostics)
-                });
-
-            assert_eq!(
-                ts5024.start,
-                source.find(value).expect("test value") as u32,
-                "TS5024 for {key} should point at the invalid value"
-            );
-
-            match key {
-                "include" => assert!(parsed.config.include.is_none()),
-                "exclude" => assert!(parsed.config.exclude.is_none()),
-                "references" => assert!(parsed.config.references.is_none()),
-                _ => unreachable!(),
-            }
-            assert_eq!(parsed.config.files, Some(vec!["a.ts".to_string()]));
-        }
-    }
-
-    #[test]
-    fn test_tsconfig_references_parsed() {
-        let source = r#"{
-            "compilerOptions": { "composite": true },
-            "references": [
-                { "path": "./packages/core" },
-                { "path": "./packages/utils", "prepend": true }
-            ]
-        }"#;
-        let parsed = parse_tsconfig_with_diagnostics(source, "tsconfig.json").unwrap();
-        let refs = parsed.config.references.expect("should have references");
-        assert_eq!(refs.len(), 2);
-        assert_eq!(refs[0].path, "./packages/core");
-        assert!(!refs[0].prepend);
-        assert_eq!(refs[1].path, "./packages/utils");
-        assert!(refs[1].prepend);
-    }
-
-    #[test]
-    fn test_extract_lib_references_normalizes_and_stops_at_first_code_line() {
-        let source = r#"
-            // regular comment
-            /// <reference lib="ES2015" />
-            /// <reference lib='lib.dom' />
-
-            const x = 1;
-            /// <reference lib="esnext" />
-        "#;
-
-        assert_eq!(
-            extract_lib_references(source),
-            vec!["es2015".to_string(), "dom".to_string()]
-        );
-    }
-
-    #[test]
-    fn test_extract_lib_references_skips_block_comments_and_ignores_embedded_lib_text() {
-        let source = r#"
-            /*
-             * /// <reference lib="es2017" />
-             */
-            /// <reference lib="es2020" />
-            /// not really a lib directive
-        "#;
-
-        assert_eq!(extract_lib_references(source), vec!["es2020".to_string()]);
-    }
-
-    #[test]
-    fn test_extract_lib_references_with_positions_anchors_at_value_start() {
-        // `/// <reference lib="notalib" />` — `notalib` starts at byte 20,
-        // matching tsc's TS2726 anchor (column 21 in 1-indexed terms).
-        let source = "/// <reference lib=\"notalib\" />\nlet x = 1;\n";
-        let refs = extract_lib_references_with_positions(source);
-        assert_eq!(refs.len(), 1, "should capture exactly one reference");
-        assert_eq!(refs[0].raw, "notalib");
-        assert_eq!(refs[0].start, 20);
-        assert_eq!(refs[0].length, 7);
-    }
-
-    #[test]
-    fn test_extract_lib_references_with_positions_handles_empty_value() {
-        let source = "/// <reference lib=\"\" />\n";
-        let refs = extract_lib_references_with_positions(source);
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].raw, "");
-        assert_eq!(refs[0].start, 20);
-        assert_eq!(refs[0].length, 0);
-    }
-
-    #[test]
-    fn test_extract_lib_references_with_positions_tracks_offset_across_lines() {
-        let source = "// header\n\n/// <reference lib=\"dom\" />\n";
-        let refs = extract_lib_references_with_positions(source);
-        assert_eq!(refs.len(), 1);
-        assert_eq!(refs[0].raw, "dom");
-        // "// header\n" = 10 bytes, blank line "\n" = 1 byte, then 11 + 20 = 31.
-        assert_eq!(refs[0].start, 31);
-        assert_eq!(refs[0].length, 3);
-    }
-
-    #[test]
-    fn test_is_known_lib_name_accepts_canonical_and_normalized_forms() {
-        // Canonical names from the TS lib catalog.
-        assert!(is_known_lib_name("es2015"));
-        assert!(is_known_lib_name("dom"));
-        // Normalization: leading `lib.` prefix and case-insensitive match.
-        assert!(is_known_lib_name("lib.es2015"));
-        assert!(is_known_lib_name("ES2015"));
-    }
-
-    #[test]
-    fn test_is_known_lib_name_rejects_empty_and_unknown() {
-        assert!(!is_known_lib_name(""));
-        assert!(!is_known_lib_name("   "));
-        assert!(!is_known_lib_name("notalib"));
-    }
-
-    #[test]
-    fn test_strip_jsonc_preserves_comment_like_text_inside_strings() {
-        let input = r#"{
-  // line comment
-  "url": "https://example.test/*keep*/",
-  "text": "// still text",
-  /* block
-     comment */
-  "value": 1
-}"#;
-
-        let stripped = strip_jsonc(input);
-        assert!(stripped.contains(r#""url": "https://example.test/*keep*/""#));
-        assert!(stripped.contains(r#""text": "// still text""#));
-        assert!(stripped.contains(r#""value": 1"#));
-        assert!(!stripped.contains("line comment"));
-        assert!(!stripped.contains("block"));
-    }
-
-    #[test]
-    fn test_normalize_jsonc_removes_comments_and_trailing_commas() {
-        let input = r#"{
-  // line comment
-  "url": "https://example.test/*keep*/",
-  "items": [
-    "a",
-  ],
-}"#;
-
-        let normalized = normalize_jsonc(input);
-        let value: serde_json::Value =
-            serde_json::from_str(&normalized).expect("normalized JSONC should parse");
-        assert_eq!(value["url"], "https://example.test/*keep*/");
-        assert_eq!(value["items"][0], "a");
-    }
-
-    #[test]
-    fn test_default_and_core_lib_names_cover_newer_targets() {
-        assert_eq!(default_lib_name_for_target(ScriptTarget::ES5), "lib");
-        assert_eq!(default_lib_name_for_target(ScriptTarget::ES2015), "es6");
-        assert_eq!(
-            default_lib_name_for_target(ScriptTarget::ES2025),
-            "esnext.full"
-        );
-        assert_eq!(
-            default_lib_name_for_target(ScriptTarget::ESNext),
-            "esnext.full"
-        );
-
-        assert_eq!(core_lib_name_for_target(ScriptTarget::ES3), "es5");
-        assert_eq!(core_lib_name_for_target(ScriptTarget::ES2025), "esnext");
-        assert_eq!(core_lib_name_for_target(ScriptTarget::ESNext), "esnext");
-    }
-
-    #[test]
-    fn parse_script_target_rejects_comma_separated_values() {
-        assert!(parse_script_target("ES5, ES2015").is_err());
-        assert!(parse_script_target("es2015, es2017").is_err());
-        assert!(parse_script_target("esnext, es2022").is_err());
-        // Single value should still work
-        assert_eq!(parse_script_target("ES5").unwrap(), ScriptTarget::ES5);
-        assert_eq!(parse_script_target("es2020").unwrap(), ScriptTarget::ES2020);
-    }
-
-    #[test]
-    fn resolve_lib_files_strict_errors_on_unknown_user_lib() {
-        // User-supplied compilerOptions.lib must error on unknown names so users
-        // catch typos. Use a name that cannot exist as an alias.
-        let err = resolve_lib_files_with_options(&["definitely.not.a.real.lib".to_string()], false)
-            .expect_err("expected unknown user lib to error");
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("unsupported compilerOptions.lib"),
-            "unexpected error message: {msg}"
-        );
-    }
-
-    #[test]
-    fn resolve_lib_files_lenient_skips_unknown_transitive_libs() {
-        // Source-file `/// <reference lib="..." />` directives should be
-        // tolerated when the lib name no longer exists. Mix a real lib with
-        // a long-renamed one to confirm the real one still resolves and the
-        // missing one is silently dropped instead of erroring.
-        let paths = resolve_lib_files_with_options_transitive(
-            &[
-                "es2018.asynciterable".to_string(),
-                "esnext.asynciterable".to_string(),
-            ],
-            false,
-        )
-        .expect("transitive resolver should not error on unknown names");
-        let names: Vec<String> = paths
-            .iter()
-            .filter_map(|p| p.file_name().and_then(|s| s.to_str()).map(String::from))
-            .collect();
-        assert!(
-            names.iter().any(|n| n.contains("es2018.asynciterable")),
-            "expected es2018.asynciterable in resolved set, got {names:?}"
-        );
-    }
-
-    #[test]
-    fn embedded_esnext_collection_follows_es2025_collection() {
-        let paths = resolve_lib_files_from_embedded(&["esnext.collection".to_string()], true)
-            .expect("embedded esnext.collection should resolve");
-        let names: Vec<String> = paths
-            .iter()
-            .filter_map(|p| p.file_name().and_then(|s| s.to_str()).map(String::from))
-            .collect();
-        assert!(
-            names.iter().any(|n| n == "es2025.collection.d.ts"),
-            "expected es2025.collection in resolved set, got {names:?}"
-        );
-        assert!(
-            names.iter().any(|n| n == "esnext.collection.d.ts"),
-            "expected esnext.collection in resolved set, got {names:?}"
-        );
-    }
-
-    /// Issue #3589 — base configs reached via `extends` must be validated by
-    /// the diagnostic loader so TS5024 surfaces on the base file rather than
-    /// being silently coerced through the non-diagnostic path.
-    #[test]
-    fn test_extends_base_invalid_boolean_string_emits_ts5024_anchored_at_base() {
-        let temp = tempdir().expect("create temp dir");
-        let base_path = temp.path().join("base.json");
-        std::fs::write(
-            &base_path,
-            r#"{
-  "compilerOptions": {
-    "noUncheckedIndexedAccess": "true"
-  }
-}"#,
-        )
-        .expect("write base");
-
-        let child_path = temp.path().join("tsconfig.json");
-        std::fs::write(
-            &child_path,
-            r#"{
-  "extends": "./base.json",
-  "files": ["a.ts"]
-}"#,
-        )
-        .expect("write child");
-
-        let parsed = load_tsconfig_with_diagnostics(&child_path).expect("load child");
-        let ts5024: Vec<&Diagnostic> = parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == diagnostic_codes::COMPILER_OPTION_REQUIRES_A_VALUE_OF_TYPE)
-            .collect();
-        assert!(
-            !ts5024.is_empty(),
-            "expected TS5024 from inherited base, got: {:?}",
-            parsed.diagnostics
-        );
-        let base_diag = ts5024
-            .iter()
-            .find(|d| d.message_text.contains("noUncheckedIndexedAccess"))
-            .expect("TS5024 for noUncheckedIndexedAccess");
-        // Normalize both sides to the canonical filename so platform path
-        // separators and intermediate `./` segments don't break the assert.
-        let base_file_name = std::path::Path::new(&base_diag.file)
-            .file_name()
-            .and_then(|s| s.to_str())
-            .unwrap_or("");
-        assert_eq!(
-            base_file_name, "base.json",
-            "TS5024 must anchor at the base file, got file={:?}",
-            base_diag.file
-        );
-        assert!(
-            !base_diag.file.contains("tsconfig.json"),
-            "TS5024 anchor must point at base.json, not the child tsconfig.json: {:?}",
-            base_diag.file
-        );
-        // The invalid coerced value must NOT survive into the merged config.
-        let opts = parsed
-            .config
-            .compiler_options
-            .as_ref()
-            .expect("merged compiler options");
-        assert_eq!(
-            opts.no_unchecked_indexed_access, None,
-            "invalidly-typed base option must be removed before merge"
-        );
-    }
-
-    /// Issue #3589 — the rule is structural, not keyed on a specific option.
-    /// Re-check with a different option (`allowJs`) to lock the parity for the
-    /// whole boolean-coercion family rather than one option name.
-    #[test]
-    fn test_extends_base_invalid_allowjs_string_emits_ts5024_anchored_at_base() {
-        let temp = tempdir().expect("create temp dir");
-        let base_path = temp.path().join("base.json");
-        std::fs::write(&base_path, r#"{"compilerOptions":{"allowJs":"true"}}"#)
-            .expect("write base");
-
-        let child_path = temp.path().join("tsconfig.json");
-        std::fs::write(&child_path, r#"{"extends":"./base.json","files":["a.ts"]}"#)
-            .expect("write child");
-
-        let parsed = load_tsconfig_with_diagnostics(&child_path).expect("load child");
-        let _ = base_path; // path is captured for inspection on failure
-        assert!(
-            parsed.diagnostics.iter().any(|d| {
-                d.code == diagnostic_codes::COMPILER_OPTION_REQUIRES_A_VALUE_OF_TYPE
-                    && std::path::Path::new(&d.file)
-                        .file_name()
-                        .and_then(|s| s.to_str())
-                        == Some("base.json")
-                    && d.message_text.contains("allowJs")
-            }),
-            "expected TS5024 anchored at base for allowJs, got: {:?}",
-            parsed.diagnostics
-        );
-    }
-
-    /// A valid base config must not produce spurious TS5024 just because the
-    /// child loader now recurses through the diagnostic path. Regression guard
-    /// for the happy path of #3589's fix.
-    #[test]
-    fn test_extends_base_valid_options_emit_no_ts5024() {
-        let temp = tempdir().expect("create temp dir");
-        let base_path = temp.path().join("base.json");
-        std::fs::write(
-            &base_path,
-            r#"{"compilerOptions":{"strict":true,"noUncheckedIndexedAccess":true}}"#,
-        )
-        .expect("write base");
-
-        let child_path = temp.path().join("tsconfig.json");
-        std::fs::write(&child_path, r#"{"extends":"./base.json","files":["a.ts"]}"#)
-            .expect("write child");
-
-        let parsed = load_tsconfig_with_diagnostics(&child_path).expect("load child");
-        assert!(
-            !parsed
-                .diagnostics
-                .iter()
-                .any(|d| d.code == diagnostic_codes::COMPILER_OPTION_REQUIRES_A_VALUE_OF_TYPE),
-            "no TS5024 expected for valid base, got: {:?}",
-            parsed.diagnostics
-        );
-        let opts = parsed
-            .config
-            .compiler_options
-            .as_ref()
-            .expect("merged compiler options");
-        assert_eq!(opts.no_unchecked_indexed_access, Some(true));
-        assert_eq!(opts.strict, Some(true));
-    }
-    fn ts6046_for_watch_options(source: &str, key: &str) -> Vec<String> {
-        let parsed =
-            parse_tsconfig_with_diagnostics(source, "tsconfig.json").expect("tsconfig must parse");
-        parsed
-            .diagnostics
-            .iter()
-            .filter(|d| d.code == 6046 && d.message_text.contains(key))
-            .map(|d| d.message_text.clone())
-            .collect()
-    }
-
-    #[test]
-    fn ts6046_fires_for_invalid_watch_file_value() {
-        // #3591 repro A: `watchOptions.watchFile: "bad"` must surface TS6046
-        // listing the valid `--watchFile` values.
-        let source = r#"{"compilerOptions":{"noEmit":true},"watchOptions":{"watchFile":"bad"},"files":["a.ts"]}"#;
-        let diags = ts6046_for_watch_options(source, "--watchFile");
-        assert_eq!(
-            diags.len(),
-            1,
-            "expected one TS6046 for invalid watchFile, got {diags:?}"
-        );
-        assert!(
-            diags[0].contains("fixedpollinginterval"),
-            "TS6046 message must list valid watchFile values, got {diags:?}"
-        );
-    }
-
-    #[test]
-    fn ts6046_fires_for_invalid_watch_directory_value() {
-        let source =
-            r#"{"compilerOptions":{"noEmit":true},"watchOptions":{"watchDirectory":"bad"}}"#;
-        let diags = ts6046_for_watch_options(source, "--watchDirectory");
-        assert_eq!(
-            diags.len(),
-            1,
-            "expected one TS6046 for invalid watchDirectory, got {diags:?}"
-        );
-    }
-
-    #[test]
-    fn ts6046_fires_for_invalid_fallback_polling_value() {
-        let source =
-            r#"{"compilerOptions":{"noEmit":true},"watchOptions":{"fallbackPolling":"bad"}}"#;
-        let diags = ts6046_for_watch_options(source, "--fallbackPolling");
-        assert_eq!(
-            diags.len(),
-            1,
-            "expected one TS6046 for invalid fallbackPolling, got {diags:?}"
-        );
-    }
-
-    #[test]
-    fn ts6046_silent_for_valid_watch_file_value() {
-        let source =
-            r#"{"compilerOptions":{"noEmit":true},"watchOptions":{"watchFile":"useFsEvents"}}"#;
-        let diags = ts6046_for_watch_options(source, "--watchFile");
-        assert!(
-            diags.is_empty(),
-            "valid watchFile value must not emit TS6046, got {diags:?}"
-        );
-    }
-}
+mod tests;
